#440 font compose for web
