+++ resources-compose/src/jsMain/kotlin/dev/icerock/moko/resources/compose/FontResource.kt
@@ -5,10 +5,29 @@
 package dev.icerock.moko.resources.compose
 
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.remember
 import androidx.compose.ui.text.font.FontFamily
+import androidx.compose.ui.text.platform.Font
 import dev.icerock.moko.resources.FontResource
+import dev.icerock.moko.resources.compose.internal.produceByteArray
 
 @Composable
 actual fun fontFamilyResource(fontResource: FontResource): FontFamily {
-    TODO()
+    val bytes: ByteArray? by produceByteArray(url = fontResource.fileUrl)
+
+    val localBytes: ByteArray? = bytes
+
+    return remember(localBytes) {
+        if (localBytes == null) {
+            return@remember FontFamily.Default
+        }
+
+        val font = Font(
+            identity = fontResource.fontFamily,
+            data = localBytes
+        )
+
+        FontFamily(font)
+    }
 }
